change class name
